refactored python structure, added check on public folder before
