+++ python/src/app/chat_utils.py
@@ -1,0 +1,3 @@
+from .obj import SystemMessenger as __SystemMessenger
+
+announcer = __SystemMessenger(display_name="Announcement", append_allow=False, save_in_history=True)
