+++ python/src/app/emotes/__init__.py
@@ -1,1 +1,2 @@
-from .emote_handling import Emotes
+from .emote_handling import Emotes as __Emotes
+emote_handler = __Emotes(False)
