+++ python/src/config-default/sidebar.json
@@ -1,0 +1,24 @@
+{
+  "sidebarCustomItems": [
+    {
+      "header": "Info",
+      "id": "logininfo_sidebar",
+      "text": "Not logged in.",
+      "textcolor": "red",
+      "media": {
+        "mediaType": "image",
+        "mediaLink": "/public/assets/blaulicht.gif",
+        "mediaId": "logininfo_picture"
+      }
+    },
+    {
+      "header": "GitHub",
+      "link": "https://github.com/FI18-Trainees/FISocketChat",
+      "text": "Check out our Repo.",
+      "media": {
+        "mediaType": "icon",
+        "mediaLink": "fab fa-github fa-2x text-white mx-2"
+      }
+    }
+  ]
+}
